+++ packages/desktop-app/src/renderer/shared/store/atoms.ts
@@ -12,6 +12,7 @@ export * from "./file-preview-atoms";
 export * from "./flowing-atoms";
 export * from "./flowing-chat-atoms";
 export * from "./project-atoms";
+export * from "./running-sessions-atoms";
 export * from "./scheduler-atoms";
 export * from "./sse-atoms";
 export * from "./todo-atoms";